# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  user.CreateUserPayload:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.UpdateUserPayload:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: API REST para gestión de mascotas y usuarios.
  title: My Pets API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []User, total: int'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List all users (system user only)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Only a user with is_system_user=true may call this endpoint.
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: 'data: User'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: email already in use
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new user (system user only)
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: user deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data: User'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get a user by ID (system user only)
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: User'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: email already in use
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update an existing user
      tags:
      - users
...
securityDefinitions:
  CookieAuth:
    description: Cookie HttpOnly de sesión. Obtenida al hacer login o refresh. Se
      envía automáticamente por el navegador.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"
...